+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HFP AG: support handling of custom AT commands with hfp_ag_register_custom_at_command
 - HFP AG: send custom unsolicited result code with hfp_ag_send_unsolicited_result_code
 - GAP: support BIG with gap_big_create and gap_big_terminate
+- GAP: support BIG Sync gap_big_sync_create and gap_big_sync_terminate
 - port: CMake build files in all windows-* ports allow to use Visual Studio 2022
 - embedded audio: mono audio is played on both channels if HAVE_HAL_AUDIO_SINK_STEREO_ONLY is defined